1
In order to run game the library file of C++ or C language that is SFML should be download separately. Only SFML 32 bit Version works properly, The reason is even if your PC or Laptop is 64 bit ,the SFML file 64 bit will not work, as working on the SFML technology a C++/C library file won't work properly, so we use 32 bit version of it.

Summary: Shapeteroids is a classic old model type of game that we use to play in 2000 in which it basically shows a plane and enemy in galaxy just like that this game have a Circle as a plane and other shapes like triangle, star and rectangle are enemy, as based on shapes that's why it is named "SHAPETEROIDS".